requirements.txt
langchain-core
langchainhub
langchain-community
langchain-google-genai
langchain-pinecone
python-dotenv
pypdf
faiss-cpu
beautifulsoup4
streamlit
streamlit-chat